+++ internal/operator/zitadel/kinds/networking/legacycf/app/filters.go
@@ -15,7 +15,7 @@ func (a *App) EnsureFilters(domain string, filters []*cloudflare.Filter) ([]*clo
 	}
 
 	createFilters, updateFilters := getFilterToCreateAndUpdate(currentFilters, filters)
-	if len(createFilters) > 0 {
+	if createFilters != nil && len(createFilters) > 0 {
 		created, err := a.cloudflare.CreateFilters(domain, createFilters)
 		if err != nil {
 			return nil, nil, err
@@ -24,7 +24,7 @@ func (a *App) EnsureFilters(domain string, filters []*cloudflare.Filter) ([]*clo
 		result = append(result, created...)
 	}
 
-	if len(updateFilters) > 0 {
+	if updateFilters != nil && len(updateFilters) > 0 {
 		updated, err := a.cloudflare.UpdateFilters(domain, updateFilters)
 		if err != nil {
 			return nil, nil, err
@@ -34,7 +34,7 @@ func (a *App) EnsureFilters(domain string, filters []*cloudflare.Filter) ([]*clo
 	}
 
 	deleteFilters := getFilterToDelete(currentFilters, filters)
-	if len(deleteFilters) > 0 {
+	if deleteFilters != nil && len(deleteFilters) > 0 {
 		del = func() error {
 			return a.cloudflare.DeleteFilters(domain, deleteFilters)
 		}
@@ -46,17 +46,19 @@ func (a *App) EnsureFilters(domain string, filters []*cloudflare.Filter) ([]*clo
 func getFilterToDelete(currentFilters []*cloudflare.Filter, filters []*cloudflare.Filter) []string {
 	deleteFilters := make([]string, 0)
 
-	for _, currentFilter := range currentFilters {
-		desc := currentFilter.Description
-		found := false
-		for _, filter := range filters {
-			if desc == filter.Description {
-				found = true
+	if filters != nil {
+		for _, currentFilter := range currentFilters {
+			desc := currentFilter.Description
+			found := false
+			for _, filter := range filters {
+				if desc == filter.Description {
+					found = true
+				}
 			}
-		}
 
-		if found == false {
-			deleteFilters = append(deleteFilters, currentFilter.ID)
+			if found == false {
+				deleteFilters = append(deleteFilters, currentFilter.ID)
+			}
 		}
 	}
 
@@ -67,20 +69,22 @@ func getFilterToCreateAndUpdate(currentFilters []*cloudflare.Filter, filters []*
 	createFilters := make([]*cloudflare.Filter, 0)
 	updateFilters := make([]*cloudflare.Filter, 0)
 
-	for _, filter := range filters {
-		found := false
-		for _, currentFilter := range currentFilters {
-			if currentFilter.Description == filter.Description ||
-				currentFilter.Expression == filter.Expression {
+	if filters != nil {
+		for _, filter := range filters {
+			found := false
+			for _, currentFilter := range currentFilters {
+				if currentFilter.Description == filter.Description ||
+					currentFilter.Expression == filter.Expression {
 
-				filter.ID = currentFilter.ID
-				updateFilters = append(updateFilters, filter)
-				found = true
-				break
+					filter.ID = currentFilter.ID
+					updateFilters = append(updateFilters, filter)
+					found = true
+					break
+				}
 			}
-		}
-		if found == false {
-			createFilters = append(createFilters, filter)
+			if found == false {
+				createFilters = append(createFilters, filter)
+			}
 		}
 	}
